+++ packages/editor/src/lib/components/default-components/DefaultBackground.tsx
@@ -1,7 +1,7 @@
 import { ComponentType } from 'react'
 
 /** @public */
-export type TLBackgroundComponent = ComponentType<object> | null
+export type TLBackgroundComponent = ComponentType
 
 export function DefaultBackground() {
 	return <div className="tl-background" />
