+++ packages/editor/src/lib/components/default-components/DefaultCollaboratorHint.tsx
@@ -6,6 +6,7 @@ import { Box2d } from '../../primitives/Box2d'
 import { Vec2d } from '../../primitives/Vec2d'
 import { clamp } from '../../primitives/utils'
 
+/** @public */
 export type TLCollaboratorHintComponent = ComponentType<{
 	className?: string
 	point: Vec2dModel
