+++ packages/editor/src/lib/components/default-components/DefaultHandles.tsx
@@ -1,21 +1,11 @@
-import { useValue } from '@tldraw/state'
 import { ComponentType } from 'react'
-import { useEditor } from '../../hooks/useEditor'
 
+/** @public */
 export type TLHandlesComponent = ComponentType<{
 	className?: string
 	children: any
 }>
 
 export const DefaultHandles: TLHandlesComponent = ({ children }) => {
-	const editor = useEditor()
-	const shouldDisplayHandles = useValue(
-		'shouldDisplayHandles',
-		() => editor.isInAny('select.idle', 'select.pointing_handle'),
-		[editor]
-	)
-
-	if (!shouldDisplayHandles) return null
-
 	return <svg className="tl-user-handles tl-overlays__item">{children}</svg>
 }
